requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
streamlit>=1.20.0
datasets>=2.10.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.60.0
pydantic>=2.0.0
requests>=2.30.0
pandas>=2.0.0